+++ src/mongo.h
@@ -163,7 +163,7 @@ typedef struct {
 typedef struct mongo {
     mongo_host_port *primary;  /**< Primary connection info. */
     mongo_replset *replset;    /**< replset object if connected to a replica set. */
-    int sock;                  /**< Socket file descriptor. */
+    size_t sock;                  /**< Socket file descriptor. */
     int flags;                 /**< Flags on this connection object. */
     int conn_timeout_ms;       /**< Connection timeout in milliseconds. */
     int op_timeout_ms;         /**< Read and write timeout in milliseconds. */
@@ -794,7 +794,7 @@ MONGO_EXPORT int mongo_get_err(mongo* conn);
 MONGO_EXPORT int mongo_is_connected(mongo* conn);
 MONGO_EXPORT int mongo_get_op_timeout(mongo* conn);
 MONGO_EXPORT const char* mongo_get_primary(mongo* conn); /* Memory returned by this function must be freed */
-MONGO_EXPORT int mongo_get_socket(mongo* conn) ;
+MONGO_EXPORT size_t mongo_get_socket(mongo* conn) ;
 MONGO_EXPORT int mongo_get_host_count(mongo* conn);
 MONGO_EXPORT const char* mongo_get_host(mongo* conn, int i); /* Memory returned by this function must be freed */
 MONGO_EXPORT mongo_cursor* mongo_cursor_create();
